Made some changes to scripts
GameManager.cs -  Now has  a reference to the maze, max X, and max Z,
all initialized on Awake. Also has "Getters" for a Cell and the entire
maze if necessary.
*BUG IN GETCELL*
EnemyAI.cs - added direction and state. Will continue to work on it
tonight. Enemies will decide their next movement every time they reach
the center of a cell.

MazeGenerator.cs - added an x and y position to Cell, which makes it
easier to access. Made each cell be initialized to the correct x and z
during maze generation. Initialized this.theMaze equal theMaze at the
end of constructMaze(). Also made getMaze return the xSize and zSize of
the maze.