Fixed some bugs - details in description
GameManger.cs - gets the reference to the maze array and can return a
certain cell or even the entire maze to any object that requests it.

EnemyAI.cs - added basic movement, does not really work yet. Still
working on it.

MazeGenerator.cs - Added X and Z to the cell and made getMaze return the
size of the maze. Also changed all the work from being done in Start()
to being done in Awake()